feature -- support 'show processlist' in standalone mode
